% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL1_lognormal}
\alias{LL1_lognormal}
\title{Log likelihood for a log-normally distributed covariate without covariate or outcome censoring}
\usage{
LL1_lognormal(X, Y, mux, sigmax, sig, alpha, b, beta_C = NULL, C = NULL)
}
\arguments{
\item{X}{a log-normally distributed covariate}

\item{Y}{outcome}

\item{mux}{mean of the ln(X)}

\item{sigmax}{standard deviation of ln(X)}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{beta_C}{regression coefficients associated with C}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a log-normally distriubted covariate when neither the covariate or outcome are censored.
}
\examples{
LL1_lognormal(X=2, Y=3, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
