% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL1_normal}
\alias{LL1_normal}
\title{Log likelihood for a normally distributed covariate without covariate or outcome censoring}
\usage{
LL1_normal(X, Y, mux, sigmax, sig, alpha, b, C = NULL, beta_C = NULL)
}
\arguments{
\item{X}{a normally distributed covariate}

\item{Y}{outcome}

\item{mux}{mean of the normally distributed covariate}

\item{sigmax}{standard deviation of the normally distributed covariate}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}

\item{beta_C}{regression coefficients associated with C}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a normally distriubted covariate when neither the covariate or outcome are censored.
}
\examples{
LL1_normal(X=2, Y=3, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
