% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL2_lognormal}
\alias{LL2_lognormal}
\title{Log likelihood for a log-normally distributed left-censored covariate without outcome censoring}
\usage{
LL2_lognormal(Y, LOD_X, mux, sigmax, sig, alpha, b, beta_C = NULL, C = NULL)
}
\arguments{
\item{Y}{outcome}

\item{LOD_X}{the limit of detection for the log-normally distributed covariate}

\item{mux}{mean of ln(X)}

\item{sigmax}{the standard deviation of ln(X)}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{beta_C}{regression coefficients associated with C}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a log-normally distriubted covariate when the covariate is left censored and the outcome is not censored.
}
\examples{
LL2_lognormal(Y=3, LOD_X = 0.2, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
