% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL2_normal}
\alias{LL2_normal}
\title{Log likelihood for a normally distributed left-censored covariate without outcome censoring}
\usage{
LL2_normal(Y, LOD_X, mux, sigmax, sig, alpha, b, C = NULL, beta_C = NULL)
}
\arguments{
\item{Y}{outcome}

\item{LOD_X}{the limit of detection for the normally distributed covariate}

\item{mux}{mean of the normally distributed covariate}

\item{sigmax}{standard deviation of the normally distributed covariate}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}

\item{beta_C}{regression coefficients associated with C}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a normally distriubted covariate when the covariate is left censored and the outcome is not censored.
}
\examples{
LL2_normal(Y=3, LOD_X = 0.2, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
