% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL3_lognormal}
\alias{LL3_lognormal}
\title{Log likelihood for a log-normally distributed covariate without censoring when the outcome is below the limit of detection}
\usage{
LL3_lognormal(X, LOD_Y, mux, sigmax, sig, alpha, b, beta_C = NULL, C = NULL)
}
\arguments{
\item{X}{the log-normally distributed covariate}

\item{LOD_Y}{the limit of detection of the outcome}

\item{mux}{mean of ln(X)}

\item{sigmax}{standard deviation of ln(X)}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{beta_C}{regression coefficients associated with C}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a log-normally distriubted covariate when the covariate is uncensored and the outcome is left censored.
}
\examples{
LL3_lognormal(LOD_Y=1, X = 2, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
