% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL3_normal}
\alias{LL3_normal}
\title{Log likelihood for a normally distributed covariate with a left censored outcome}
\usage{
LL3_normal(X, LOD_Y, mux, sigmax, sig, alpha, b, C = NULL, beta_C = NULL)
}
\arguments{
\item{X}{a normally distributed covariate}

\item{LOD_Y}{the limit of detection for the outcome}

\item{mux}{mean of the normally distributed covariate}

\item{sigmax}{standard deviation of the normally distributed covariate}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}

\item{beta_C}{regression coefficients associated with C}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a normally distriubted covariate when the outcome is left censored but the covariate is above the limit of detection
}
\examples{
LL3_normal(LOD_Y=1, X = 2, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
