% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL4_lognormal}
\alias{LL4_lognormal}
\title{Log likelihood for a log-normally distributed covariate when both the covariate and outcome are below the limit of detection}
\usage{
LL4_lognormal(
  LOD_X,
  LOD_Y,
  mux,
  sigmax,
  sig,
  alpha,
  b,
  beta_C = NULL,
  C = NULL,
  single_LODs = F
)
}
\arguments{
\item{LOD_X}{the limit of detection of the log-normally distributed covariate}

\item{LOD_Y}{the limit of detection of the outcome}

\item{mux}{mean of ln(X)}

\item{sigmax}{standard deviation of ln(X)}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{beta_C}{regression coefficients associated with C}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a log-normally distriubted covariate when the covariate and the outcome are both left censored.
}
\examples{
LL4_lognormal(LOD_Y=1, LOD_X = 0.2, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
