% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{LL4_normal}
\alias{LL4_normal}
\title{Log likelihood for a normally distributed covariate with a left censored outcome}
\usage{
LL4_normal(
  LOD_X,
  LOD_Y,
  mux,
  sigmax,
  sig,
  alpha,
  b,
  C = NULL,
  beta_C = NULL,
  single_LODs = F
)
}
\arguments{
\item{LOD_X}{the limit of detection for the normally distributed covariate}

\item{LOD_Y}{the limit of detection for the outcome}

\item{mux}{mean of the normally distributed covariate}

\item{sigmax}{standard deviation of the normally distributed covariate}

\item{sig}{the standard deviation of the residuals}

\item{alpha}{intercept: the average value of Y when all covariates = 0}

\item{b}{slope: the average change in Y for a 1 unit change in X}

\item{C}{optional vector of other covariates (not subject to LODs) included in the model}

\item{beta_C}{regression coefficients associated with C}
}
\value{
log-likelihood
}
\description{
Calculates the log-likelihood contribution for a normally distriubted covariate when both the outcome and covariate are left censored.
}
\examples{
LL4_normal(LOD_Y=1, LOD_X = 0.2, sig = 0.5, mux = 1, sigmax = 1, alpha = 0, b = 1)

}
