% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle2}
\alias{mle2}
\title{Maximum likelihood estimation of model paramters}
\usage{
mle2(
  minuslogl,
  start = NULL,
  method = "Nelder-Mead",
  nobs,
  maxit = 1000,
  reltol = 1e-08,
  ...
)
}
\arguments{
\item{minuslogl}{a negative log likelihood function to minimize}

\item{start}{a vector of starting values for the parameters}

\item{method}{optimization method to use. See optim. Default is "Nelder-Mead".}

\item{nobs}{optional integer: the number of observations, to be used for e.g. computing BIC.}

\item{maxit}{maximum number of iterations}

\item{reltol}{relative tolerance for stopping optimization.  See optim.}

\item{...}{Further arguments to pass to optim.}
}
\value{
an object of class mle
}
\description{
Maximum likelihood estimation of model paramters
}
\examples{
x <- rnorm(100, mean = 1)
y <- 3*x + rnorm(100, mean = 0, sd=0.2)
censor_x <- ifelse(x < 0.2, 1, 0)
censor_y <- ifelse(y < 0, 1, 0)
mle2(start = c(mean(x), sd(x), sd(y), 0, 0), minuslogl = function(param){NNL(param, LOD_x = rep(0.2,100), LOD_y = rep(1,100), x = x, y = y, censor_x = censor_x, censor_y = censor_y, C=NULL, single_LODs=T, x.dist = 'normal')}, maxit = 1000)


}
