% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mle_fit}
\alias{mle_fit}
\title{Maximum Likelihood Models with Censored Outcomes and Covariates}
\usage{
mle_fit(
  y,
  x,
  censor_x,
  censor_y,
  LOD_x = NULL,
  LOD_y = NULL,
  covariates = NULL,
  x.dist = "gaussian",
  start = NULL,
  method = "Nelder-Mead",
  maxit = 5000,
  reltol = 1e-08,
  maxoptimit = 10,
  llabstol = 1e-08
)
}
\arguments{
\item{y}{a vector of potentially censored outcome values}

\item{x}{a vecotr of potentially censored covariate values}

\item{censor_x}{a vector the length of x; 0 if x is observed, 1 if censored}

\item{censor_y}{a vector the length of y; 0 if y is observed, 1 if censored}

\item{LOD_x}{a single LOD for the censored covariate or a vector of LODs (specific to each observation)}

\item{LOD_y}{a single LOD for the outcome or a vector of LODs (specific to each observation)}

\item{covariates}{an optional matrix or dataframe of additional uncensored covariates to include in the model}

\item{x.dist}{the distribution for censored predictor - either 'gaussian', 'normal' or 'lognormal'}

\item{start}{a vector of initial values for optimizer in the following order: mu_x, sigma_x, sigma_y, intercept, slope, covariates (in same order supplied in covariates.  If not supplied will default to mu_x = mean(x) if gaussian x, mean(log(x)) if lognormal x, sigma_x = sd(x) if gaussian x, sd(log(x)) if lognormal x, sigma_y = sd(y), intercept = mean(y), slope = 0, covariates = 0}

\item{method}{Optimization method to use. See optim.}

\item{maxit}{integer. Maximum number of iterations for each optmization run.}

\item{reltol}{relative tolerance for stopping optimization.  See optim.}

\item{maxoptimit}{maximum number of opitimization iterations}

\item{llabstol}{absolute tolerance (change in log likelihood) for declaring convergence}
}
\value{
an object of class mle
}
\description{
Calculates model parameters when either the covariate, outcome or both are subject to a lower limit of detection.
}
\examples{
x <- rnorm(100, mean = 1)
y <- 3*x + rnorm(100, mean = 0, sd=0.2)
censor_x <- ifelse(x < 0.2, 1, 0)
censor_y <- ifelse(y < 0, 1, 0)
mle_fit(y=y, x=x,
       censor_x = censor_x, censor_y = censor_y,
       LOD_x = rep(0.2,100), LOD_y = rep(1,100),
       x.dist = 'normal')

}
