% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{NNL}
\alias{NNL}
\title{Negative Log Likelihood}
\usage{
NNL(
  param,
  LOD_x,
  LOD_y,
  x,
  y,
  censor_x,
  censor_y,
  C = NULL,
  single_LODs = F,
  x.dist
)
}
\arguments{
\item{param}{a vector of values for the model paramters in the following order c(mux, sigmax, sig, alpha, b, beta_C)}

\item{LOD_x}{a single limit of detection for the censored covariate or a vector of LODs (specific to each observation)}

\item{LOD_y}{a single limit of detection for the outcome or a vector of LODs (specific to each observation)}

\item{x}{a vector of potentially left censored predictor values}

\item{y}{a vecotr of potentially left censored outcome values}

\item{censor_x}{a vector the length of x; 0 if x is observed, 1 if censored}

\item{censor_y}{a vector the length of y; 0 if y is observed, 1 if censored}

\item{C}{optional matrix or data frame of other covariates (not subject to LODs) included in the model}

\item{single_LODs}{do observations have varying limits of detection.  Logical.}

\item{x.dist}{distribution for the X covariate that is subject to a limit of detection.  'normal', 'gaussian' or 'lognormal'.}
}
\value{
negative log-likelihood of a model for the data
}
\description{
Calculates the negative log-likelihood of a model for the data
}
\examples{
x <- rnorm(100, mean = 1)
y <- 3*x + rnorm(100, mean = 0, sd=0.2)
censor_x <- ifelse(x < 0.2, 1, 0)
censor_y <- ifelse(y < 0, 1, 0)
NNL(param = c(1, 1, 0.2, alpha = 0, b=3), LOD_x = rep(0.2,100), LOD_y = rep(1,100), x = x, y = y, censor_x = censor_x, censor_y = censor_y, C=NULL, single_LODs=T, x.dist = 'normal')

}
